Abstract DocxTaskWriter into TaskWriter
